+++ tests/Unit/Liker/Exceptions/LikerNotDefinedTest.php
@@ -9,27 +9,27 @@
  * file that was distributed with this source code.
  */
 
-namespace Cog\Tests\Laravel\Likeable\Unit;
+namespace Cog\Tests\Laravel\Likeable\Unit\Liker\Exceptions;
 
-use Cog\Contracts\Likeable\Like\Exceptions\LikerNotDefinedException;
+use \Cog\Contracts\Likeable\Liker\Exceptions\LikerNotDefined;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\Entity;
 use Cog\Tests\Laravel\Likeable\Stubs\Models\User;
 use Cog\Tests\Laravel\Likeable\TestCase;
 use Illuminate\Foundation\Testing\DatabaseTransactions;
 
 /**
- * Class LikerNotDefinedExceptionTest.
+ * Class LikerNotDefinedTest.
  *
- * @package Cog\Tests\Laravel\Likeable\Unit\Exceptions
+ * @package Cog\Tests\Laravel\Likeable\Unit\Liker\Exceptions
  */
-class LikerNotDefinedExceptionTest extends TestCase
+class LikerNotDefinedTest extends TestCase
 {
     use DatabaseTransactions;
 
     /** @test */
     public function it_can_throw_exception_if_not_authenticated_on_like()
     {
-        $this->expectException(LikerNotDefinedException::class);
+        $this->expectException(LikerNotDefined::class);
 
         $entity = factory(Entity::class)->create();
 
@@ -39,7 +39,7 @@ class LikerNotDefinedExceptionTest extends TestCase
     /** @test */
     public function it_can_throw_exception_if_authenticated_but_passed_zero_on_like()
     {
-        $this->expectException(LikerNotDefinedException::class);
+        $this->expectException(LikerNotDefined::class);
 
         $entity = factory(Entity::class)->create();
         $user = factory(User::class)->create();
@@ -51,7 +51,7 @@ class LikerNotDefinedExceptionTest extends TestCase
     /** @test */
     public function it_can_throw_exception_if_not_authenticated_on_unlike()
     {
-        $this->expectException(LikerNotDefinedException::class);
+        $this->expectException(LikerNotDefined::class);
 
         $entity = factory(Entity::class)->create();
 
@@ -61,7 +61,7 @@ class LikerNotDefinedExceptionTest extends TestCase
     /** @test */
     public function it_can_throw_exception_if_authenticated_but_passed_zero_on_unlike()
     {
-        $this->expectException(LikerNotDefinedException::class);
+        $this->expectException(LikerNotDefined::class);
 
         $entity = factory(Entity::class)->create();
         $user = factory(User::class)->create();
@@ -73,7 +73,7 @@ class LikerNotDefinedExceptionTest extends TestCase
     /** @test */
     public function it_can_throw_exception_if_not_authenticated_on_where_liked_by()
     {
-        $this->expectException(LikerNotDefinedException::class);
+        $this->expectException(LikerNotDefined::class);
 
         Entity::whereLikedBy();
     }
@@ -81,7 +81,7 @@ class LikerNotDefinedExceptionTest extends TestCase
     /** @test */
     public function it_can_throw_exception_if_authenticated_but_passed_zero_on_where_liked_by()
     {
-        $this->expectException(LikerNotDefinedException::class);
+        $this->expectException(LikerNotDefined::class);
 
         $user = factory(User::class)->create();
         $this->actingAs($user);
